+++ dlls/Game/model/Tetrimino.h
@@ -60,7 +60,7 @@ namespace model
 		{
 			for ( uint8_t i = 0; i != ::model::tetrimino::BLOCKS_A_TETRIMINO*::model::tetrimino::BLOCKS_A_TETRIMINO; ++i )
 			{
-				if ( (mPossibleRotations[static_cast<uint8_t>(mRotationID)]>>i) & 1u ) //궁금: msb 0으로 채워질까 1일까?
+				if ( (mPossibleRotations[static_cast<int>(mRotationID)]>>i) & 1u ) //궁금: msb 0으로 채워질까 1일까?
 				{
 					// Coordinate transformation
 					const sf::Vector2< int8_t > localPos( i%model::tetrimino::BLOCKS_A_TETRIMINO, i/model::tetrimino::BLOCKS_A_TETRIMINO );
@@ -84,7 +84,7 @@ namespace model
 		// Current blocks within their own local space.
 		inline LocalSpace blocks( ) const
 		{
-			return mPossibleRotations[ static_cast< uint8_t >( mRotationID ) ];
+			return mPossibleRotations[ static_cast<int>(mRotationID) ];
 		}
 		// Returns true when colliding with the floor or another tetrimino.
 		inline bool down( const std::array< std::array<::model::Cell,::model::stage::GRID_WIDTH>, ::model::stage::GRID_HEIGHT >& grid, const uint8_t diff = 1u )
@@ -143,7 +143,7 @@ namespace model
 		float mCellSize_;
 		sf::Vector2f mOrigin_;
 		sf::RectangleShape mCellShape;
-		LocalSpace mPossibleRotations[ static_cast< uint8_t >( Rotation::NULL_MAX ) ];
-		static sf::Vector2<int8_t> Test[ static_cast< uint8_t >( Rotation::NULL_MAX ) ][ 4 ];
+		LocalSpace mPossibleRotations[ static_cast<int>(Rotation::NULL_MAX) ];
+		static sf::Vector2<int8_t> Test[ static_cast<int>(Rotation::NULL_MAX) ][ 4 ];
 	};
 }
